// pfplugin/packages/html-comment/src/slate-yjs/core/types/types.ts
import type Y from 'yjs';
import type { SlateElement, SlateEditor, SlateNode } from '@wangeditor/editor';

export type DeltaAttributes = {
  retain: number;
  attributes: Record<string, unknown>;
};

export type DeltaRetain = { retain: number };
export type DeltaDelete = { delete: number };
export type DeltaInsert = {
  insert: string | Y.XmlText;
  attributes?: Record<string, unknown>;
};

export type InsertDelta = Array<DeltaInsert>;
export type Delta = Array<
  DeltaRetain | DeltaDelete | DeltaInsert | DeltaAttributes
>;

export type TextRange = { start: number; end: number };

export type HistoryStackItem = {
  meta: Map<string, unknown>;
};

export type YTarget = {
  // TextRange in the yParent mapping to the slateTarget (or position to insert)
  textRange: TextRange;

  // Y.XmlText containing the slate node
  yParent: Y.XmlText;

  // Slate element mapping to the yParent
  slateParent: SlateElement | SlateEditor;

  // If the target points to a slate element, Y.XmlText representing the target.
  // If it points to a text (or position to insert), this will be undefined.
  yTarget?: Y.XmlText;

  // Slate node represented by the textRange, won't be set if position is insert.
  slateTarget?: SlateNode;

  // InsertDelta representing the slateTarget
  targetDelta: InsertDelta;
};

export type RelativeRange = {
  anchor: Y.RelativePosition;
  focus: Y.RelativePosition;
};